Fix rendering XML.
Not sure I understand this one, but the highlight function does not work
if it is passed the jQuery wrapper object, and it does not work either
if it is passed this directly.  The only way it works is by calling the
jQuery function on this, but passing this itself.