New browsing system.
The new system is based on each path being associated to a root and
a path separator.  It is then possible to support various URI schemes.
Common schemes (linke starting with '/', 'http://' or 'urn:', or just
containing '/' as a separator) are supported.  It is possible to
explicitly provide roots and separator when using other schemes.

The old system still exists, and still needs to be completely removed.