New component to select an exec target in a form.
An exec target is either a database or an appserver.  All of them
available in the instance are displayed (in 5 different dropdown
lists: all databases, HTTP servers, XDBC servers, ODBC servers and
WebDAV servers.)

The value is used to eval/invoke some code (or to put it simply: to
"execute" it.)  If a database is passed, it is used as the content
database for execution.  If an appserver is passed, both its content
and modules databases are used as such for execution.